pyproject.toml
[project]
name = "computational-thinking-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "dnspython>=2.8.0",
    "firebase-admin>=7.1.0",
    "flask>=3.1.2",
    "google-auth-oauthlib>=1.2.3",
    "google-cloud-firestore>=2.21.0",
    "google-genai>=1.52.0",
    "numpy>=2.3.5",
    "playwright>=1.56.0",
    "pyqt6>=6.10.0",
    "pysocket>=0.35",
    "python-dotenv>=1.2.1",
    "python-whois>=0.9.6",
    "requests>=2.32.5",
    "selenium>=4.0.0",
    "tldextract>=5.3.0",
    "urllib3>=2.5.0",
    "webdriver-manager>=4.0.0",
    "groq>=0.1.0",
]
